requirements.txt
Django==2.0.6
django-crispy-forms==1.7.2
django-js-asset==1.1.0
django-ckeditor==5.5.0
